Utilize different display logic depending on version
